Create syntactic_indexing module, extract job store into it
