+++ cmd/syntactic-code-intel-worker/shared/indexing_worker.go
@@ -2,20 +2,24 @@ package shared
 
 import (
 	"context"
-	"github.com/sourcegraph/log"
 	"time"
 
+	"github.com/sourcegraph/log"
+
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/syntactic_indexing/jobstore"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/workerutil"
 	"github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker"
-	dbworkerstore "github.com/sourcegraph/sourcegraph/internal/workerutil/dbworker/store"
 )
 
-func NewIndexingWorker(ctx context.Context, observationCtx *observation.Context, workerStore dbworkerstore.Store[*SyntacticIndexingJob], config IndexingWorkerConfig) *workerutil.Worker[*SyntacticIndexingJob] {
+func NewIndexingWorker(ctx context.Context,
+	observationCtx *observation.Context,
+	jobStore jobstore.SyntacticIndexingJobStore,
+	config IndexingWorkerConfig) *workerutil.Worker[*jobstore.SyntacticIndexingJob] {
 
 	name := "syntactic_code_intel_indexing_worker"
 
-	return dbworker.NewWorker[*SyntacticIndexingJob](ctx, workerStore, &indexingHandler{}, workerutil.WorkerOptions{
+	return dbworker.NewWorker(ctx, jobStore.DBWorkerStore(), &indexingHandler{}, workerutil.WorkerOptions{
 		Name:                 name,
 		Interval:             config.PollInterval,
 		HeartbeatInterval:    10 * time.Second,
@@ -28,9 +32,9 @@ func NewIndexingWorker(ctx context.Context, observationCtx *observation.Context,
 
 type indexingHandler struct{}
 
-var _ workerutil.Handler[*SyntacticIndexingJob] = &indexingHandler{}
+var _ workerutil.Handler[*jobstore.SyntacticIndexingJob] = &indexingHandler{}
 
-func (i indexingHandler) Handle(ctx context.Context, logger log.Logger, record *SyntacticIndexingJob) error {
+func (i indexingHandler) Handle(ctx context.Context, logger log.Logger, record *jobstore.SyntacticIndexingJob) error {
 	logger.Info("Stub indexing worker handling record",
 		log.Int("id", record.ID),
 		log.String("repository name", record.RepositoryName),
