+++ internal/codeintel/syntactic_indexing/jobstore/store_helpers_test.go
@@ -1,4 +1,4 @@
-package shared
+package jobstore
 
 import (
 	"context"
